Projects/ATS-System/requirement.txt
streamlit
google-generativeai
python-dotenv 
pdf2image 